cronjobs/requirements.txt
schedule==0.4.3
pymongo==3.5.1
requests==2.18.4
redis==2.10.5
psutil==5.0.1
pystache==0.5.4
pandas==0.17.0
sparkpost==1.3.6
sklearn==0.0
python-dotenv==0.9.1
